requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aioresponses==0.7.6
aiosignal==1.3.1
alembic==1.13.3
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
backoff==2.2.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.0.1
defusedxml==0.7.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dspy==2.5.3
fastapi==0.115.2
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
groq==0.11.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
importlib_metadata==8.5.0
inference-sdk==0.9.17
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
litellm==1.48.19
magicattr==0.1.6
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==3.0.0
marshmallow==3.22.0
matplotlib==3.9.2
mdurl==0.1.2
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.9.2
numpy==2.1.1
openai==1.51.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
optuna==4.0.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
protobuf==5.28.2
py-cpuinfo==9.0.0
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
scipy==1.14.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.40.0
streamlit==1.39.0
structlog==24.4.0
supervision==0.23.0
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.20.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
watchdog==5.0.3
xxhash==3.5.0
yarl==1.13.1
zipp==3.20.2
